scripts/python/requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
